chore: Use consistent fs structure
