+++ src/shared/ui/components/tabs/index.tsx
@@ -1,5 +1,5 @@
-import React, { useCallback } from "react";
 import { Tabs as MuiTabs, Tab, Box, TabProps } from "@mui/material";
+import React, { useCallback } from "react";
 
 interface Props {
     titles: string[] | JSX.Element[];
@@ -13,6 +13,34 @@ interface Props {
 
 const TabPanel = ({ children, value, index }: { children: React.ReactNode; value: number; index: number }) => (
     <div hidden={value !== index}>{value === index && children}</div>
+);
+
+const CustomTab = (props: TabProps) => (
+    <Tab
+        {...props}
+        sx={{
+            width: 116,
+            height: 44,
+            ml: "19px",
+            mb: "35px",
+            mt: "20px",
+            fontFamily: "Titillium Web",
+            fontWeight: "bold",
+            textTransform: "none",
+            borderRadius: 1,
+            background: "transparent",
+            color: "#e0e0e0",
+            fontSize: "16px",
+            border: "1px solid #e0e0e0 ",
+            opacity: 1,
+
+            "&.Mui-selected": {
+                background: "#A4BAB8",
+                color: "#2a2a2a",
+                borderColor: "#000000",
+            },
+        }}
+    />
 );
 
 export const Tabs = ({ titles, contents, customCurrentTab, customOnChange, tabPadding, CustomTab }: Props) => {
@@ -80,3 +108,11 @@ export const Tabs = ({ titles, contents, customCurrentTab, customOnChange, tabPa
         </Box>
     );
 };
+
+export const PageTabs = ({ contents }: { contents: JSX.Element[] }) => (
+    <Tabs
+        titles={["Multicall", "DAO"]}
+        contents={contents}
+        CustomTab={CustomTab}
+    />
+);
